+++ test/emptiable_list_test.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:emptiable_list/emptiable_list.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
@@ -6,14 +8,14 @@ const ANIMATION_DURATION = const Duration(milliseconds: 1);
 
 class DummyWidget extends StatelessWidget {
   final Stream<List<String>> items;
-  final bool defaultPlaceholder;
+  static final initializingPlaceholder = Text('Loading');
   static final placeholder = Text('Empty list placeholder');
   static final list = ListView(children: <Widget>[
     ListTile(title: Text('Item 1')),
     ListTile(title: Text('Item 2'))
   ]);
 
-  DummyWidget({@required this.items, this.defaultPlaceholder = false});
+  DummyWidget({@required this.items});
 
   @override
   Widget build(BuildContext context) {
@@ -22,7 +24,8 @@ class DummyWidget extends StatelessWidget {
         body: EmptiableList(
           listStream: items,
           list: list,
-          placeholder: defaultPlaceholder ? null : placeholder,
+          placeholder: placeholder,
+          initializingPlaceholder: initializingPlaceholder,
           transitionDuration: ANIMATION_DURATION,
         ),
       ),
@@ -40,8 +43,11 @@ void main() {
     testWidgets('displays placeholder when list, coming from stream, is empty',
         (WidgetTester tester) async {
       await tester.pumpWidget(DummyWidget(items: Stream.value([])));
+      await tester.pump(ANIMATION_DURATION);
+      final initializationPlaceholderOpacity = findOpacityOf(DummyWidget.initializingPlaceholder, tester);
       final placeholderOpacity = findOpacityOf(DummyWidget.placeholder, tester);
       final listOpacity = findOpacityOf(DummyWidget.list, tester);
+      expect(initializationPlaceholderOpacity.opacity, 0);
       expect(placeholderOpacity.opacity, 1);
       expect(listOpacity.opacity, 0);
     });
@@ -49,26 +55,25 @@ void main() {
         'displays list widget when list, coming from stream, is not empty',
         (WidgetTester tester) async {
       await tester.pumpWidget(DummyWidget(items: Stream.value(['Item'])));
-      await tester.pumpAndSettle(ANIMATION_DURATION);
+      await tester.pump(ANIMATION_DURATION);
+      final initializationPlaceholderOpacity = findOpacityOf(DummyWidget.initializingPlaceholder, tester);
       final placeholderOpacity = findOpacityOf(DummyWidget.placeholder, tester);
       final listOpacity = findOpacityOf(DummyWidget.list, tester);
+      expect(initializationPlaceholderOpacity.opacity, 0);
       expect(placeholderOpacity.opacity, 0);
       expect(listOpacity.opacity, 1);
     });
-    testWidgets(
-        'displays nothing when list, coming from stream, is empty and the placeholder widget was not specified',
-        (WidgetTester tester) async {
-      await tester.pumpWidget(
-          DummyWidget(items: Stream.value([]), defaultPlaceholder: true));
-      await tester.pumpAndSettle(ANIMATION_DURATION);
-      final placeholderFinder = find.byWidget(DummyWidget.placeholder);
+    testWidgets('displays initialization placeholder while waiting for an emit form a stream', (WidgetTester tester) async {
+      final controller = StreamController<List<String>>();
+      await tester.pumpWidget(DummyWidget(items: controller.stream));
+      await tester.pump(ANIMATION_DURATION);
+      final initializationPlaceholderOpacity = findOpacityOf(DummyWidget.initializingPlaceholder, tester);
+      final placeholderOpacity = findOpacityOf(DummyWidget.placeholder, tester);
       final listOpacity = findOpacityOf(DummyWidget.list, tester);
-      final sizedBoxFinder = find.byWidget(EMPTY_WIDGET);
-      final sizedBoxOpacity = findOpacityOf(EMPTY_WIDGET, tester);
-      expect(placeholderFinder, findsNothing);
-      expect(sizedBoxFinder, findsOneWidget);
+      expect(initializationPlaceholderOpacity.opacity, 1);
+      expect(placeholderOpacity.opacity, 0);
       expect(listOpacity.opacity, 0);
-      expect(sizedBoxOpacity.opacity, 1);
+      await controller.close();
     });
   });
 }