+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [1.1.0]
+
+* While waiting for a first version of a list from a listStream,
+EmptiableList will display initializingPlaceholder.
+
 ## [1.0.1]
 
 * Added documentation.